Testing complete
After some testing I have found some bugs but the AI is working in a rough way. Just need to tweak the AI variables so they are not to op and make it so that the AI will stop moving once it is in attack range and not attack and move if we don't want that.